requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
autopep8==2.0.0
Brotli==1.0.9
certifi==2022.5.18
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cssselect2==0.7.0
daphne==3.0.2
defusedxml==0.7.1
Django==4.1.3
django-cors-headers==3.12.0
django-environ==0.8.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
fonttools==4.38.0
future==0.18.2
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
msgpack==1.0.3
oauthlib==3.2.0
oscrypto==1.3.0
packaging==21.3
pdfkit==1.0.0
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.5.0
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyphen==0.13.2
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.64.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
weasyprint==57.1
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.8
zope.interface==5.4.0
zopfli==0.2.2
